--- src/main/java/conflux/web3j/utils/ToStringUtils.java
@@ -1,11 +1,0 @@
-package conflux.web3j.utils;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-
-public class ToStringUtils {
-    private static final ObjectMapper objectMapper = new ObjectMapper();
-
-    public static ObjectMapper getObjectMapper() {
-        return objectMapper;
-    }
-}
