+++ src/main/java/conflux/web3j/utils/Utils.java
@@ -1,0 +1,21 @@
+package conflux.web3j.utils;
+
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.ObjectMapper;
+
+public class Utils {
+    private static final ObjectMapper objectMapper = new ObjectMapper();
+
+    public static ObjectMapper getObjectMapper() {
+        return objectMapper;
+    }
+
+    public static String jsonStringify(Object data) {
+        try {
+            return objectMapper.writeValueAsString(data);
+        } catch (JsonProcessingException e) {
+            e.printStackTrace();
+            return "{\"className\":" + data.toString() + "}";
+        }
+    }
+}
